remove redundant long casting when setting disk provisioned IOPS and throughput in DataflowPipelineTranslator

+++ runners/google-cloud-dataflow-java/src/main/java/org/apache/beam/runners/dataflow/DataflowPipelineTranslator.java
@@ -490,12 +490,11 @@ public class DataflowPipelineTranslator {
         workerPool.setDiskSizeGb(options.getDiskSizeGb());
       }
       if (options.getDiskProvisionedIops() != null && options.getDiskProvisionedIops() > 0) {
-        workerPool.setDiskProvisionedIops((long) options.getDiskProvisionedIops());
+        workerPool.setDiskProvisionedIops(options.getDiskProvisionedIops());
       }
       if (options.getDiskProvisionedThroughputMibps() != null
           && options.getDiskProvisionedThroughputMibps() > 0) {
-        workerPool.setDiskProvisionedThroughputMibps(
-            (long) options.getDiskProvisionedThroughputMibps());
+        workerPool.setDiskProvisionedThroughputMibps(options.getDiskProvisionedThroughputMibps());
       }
       AutoscalingSettings settings = new AutoscalingSettings();
       if (options.getAutoscalingAlgorithm() != null) {
